Fix probable bug with csh in config script.
At least on my Ubuntu workstations, I get an "illegal variable"
error when running "./config" with no arguments. That is, no <arch>
options are displayed based on the setup.

This appears to trace back to the behavior of $?, which is supposed
reference a variable that has not been set yet (or something, I'm
more of a bash/dash guy, so someone please correct me if they know
better). I introduced a dummy variable when detecing cpu info and
this error seems to go away and the correct behavior is produced.

Change-Id: I2cfb32997c5e4a22a378eb3066f2540f2707320e