Fix output error when writing velnamdbin
A modest output error was added when psfgen was modified to write
velnamdbin files as a secondary option for writenamdbin. The buffer
was simply being overwritten by the velnamdbin message rather than
first being printed to stdout. This is now fixed.

To reiterate, this only affected the message sent to stdout, the
actual files were of course being written as intended. I just
happened to be checking this carefully when I noticed the message
was not appearing.

Change-Id: I7d021ebc4d1e5b095fcbfdefd02040b2a6cd0f5c