Modification of LJcorrection for vdwForceSwitching
  Testing revealed that the previous implementation of an
LJcorrection for vdwForceSwitching was based on erroneous
assumptions - the energy was always overestimated and decayed
slowly as a function of cutoff. This has now been fixed at the
expense of re-defining the potential (but NOT the forces) when
LJcorrection is on. Briefly, there is a potential shift applied to
all interacting pairs _within the cutoff distance_ (or rather, such
a shift is unapplied). The resulting potential is formally
discontinuous at the cutoff distance, but the (analytic) forces are
still continous. The virial corrections are unchanged.

This really only has serious repercussions when the energy is used
as an observable (e.g. absolute solvation free energy or enthalpy of
vaporization). Since the virial corrections are still the same,
these adjustments can be rigorously applied as post-processing.

Minor bugfix - A division by zero error occured when using alchemy
with one group and LJcorrection. This is because the average LJ
coefficients were being computed over zero atoms for the empty
alchemical group - this is now fixed and the coefficients are set
at zero.

Change-Id: Id4ba2e183726cff1e33594369fac9f2ef0352b1a