Fixed bug in CUDA kernels.
The kernels from A-P had a logic error that did not account for
the fact that, when simulating a small enough system, a patch might
be interacting with a periodic image of itself.  The problem appears
in two different CUDA kernels: buildTileListsBBKernel that reorganizes
the patch-patch interactions into tiles sized 32-by-32 and
nonbondedForceKernel that does the actual computation.  In both cases,
the A-P kernel sets bool "self" to true if a tile is interacting with
itself, without taking into account that there might be a nonzero
offset vector between the interactions.  The code is changed slightly
to set "self" only if the offset vector is zero.

Change-Id: Id787b654b3b45319c2f7e03ed2c34775b4453ee3