Documentation for constant-pH MD
This only submits changes for the user guide. A preliminary
description of the method and accompanying keywords is given. Some
experimental options are NOT listed and this is noted in the manual
as well. Since the manuscript is still in review, I will almost
certainly have to make a quick update in the near future.

Change-Id: I3756a1761bce4e6f6cd415cd27347b51daefee52