Bug fix for PME kernel in CUDA 9
Contributed by Ke Li (NVIDIA) with assistance from software QA.
Found a CUDA 9 bug in A-P's PME kernel due to incorrect masking
of synchronized warp operations. The issue is in the gather_force
kernel, where if the number of atoms being processed by a warp is
not a multiple of 4, some threads will exit the loop early.
Using the full warp mask 0xffffffff here is incorrect and will
actually cause the warp operations to hang.

Note that for the two test cases that NVIDIA had been using,
ApoA1 and STMV, both have atom counts that are multiple of 4,
which is why the bug had not initially been triggered. The bug was
finally caught when running ApoA1 on an 8-GPU platform (DGX-1).
For finding any similar bugs, a box of water turns out to be a
worthwhile test case.

Change-Id: I7cff7cc57c5b32eb586ef5f0f874ad530f322582