Bug Fix for NAMD Non-Uniform Grids
When attempting to use non-uniform grids with subgrids of generation greater than one, NAMD would fail with an error message or segmentation fault, depending on the system. This result occurred because GridforceFullSubGrid objects instantiated in the readSubgridHierarchy method did not have the poten_fp attribute set. This fix ensures that this attribute is set by the relevant constructor, and testing has indicated that it allows NAMD to successfully use non-uniform grids with subgrids of generation greater than one.

Change-Id: Ib33323387f23b1be19af11c7bdf0370ef31b53e2