Automatic detection of lonepairs
Following the behavior in CHARMM, a lonepair section of a PSF is
now automatically detected without the lonepairs keyword being set.
In practice, all of the same keywords still exist, but now default to
TRUE and are instead toggled to FALSE if no lonepairs are present.

PSF reading in Molecule will now optionally read a lone pair hosts
section when it finds the NUMLP keyword. It will disable the "lonepairs"
flag in SimParameters if no lone pair hosts are found. An error occurs
if lone pair hosts are found with "lonepairs=false" set in config file.

Change-Id: I5f08ff37646567994d0e3df35084665858932be9