Fixed DCD file writer for billion atom systems
The DCD file writer had been broken for systems of size N >= 2^30 due to
representing byte counts and some of the file seeking byte counts using
variables of size int32. This patch fixes the problem by now using
size_t for the block byte sizes and OFF_T for the file seeking.

However, the DCD file format is intrinsically broken for these larger
systems, since it requires that the byte counts of these blocks of
N-floats be stored as int32. For now, we will simply write these
overflowed values into the DCD file, preserving the expected byte count
for each frame, and VMD will set a flag for systems of N >= 2^30 to
ignore the stored value of these block byte count markers.

Change-Id: I13b0fe0573be3b4b25f3504a075fdefdc3641aca