Improve PME performance when toggling alch/LES/PairInteraction
The number of PME grids is increased for methods like alchemy and
causes a slowdown of ~30%, even if the method is toggled off at
the Tcl level (as happens in hybrid neMD/MC, such as constant pH).
This commit enables the number of PME grids to change when the MD
settings permit it.

Following the ComputeNonbondedUtil class as a template, this
introduces the new ComputePmeUtil class that caches the PME grid
related variables from SimParameters and shares that information
with ComputePme and ComputePmeMgr via inheritance. Other variables
could also be stored here in order to centralize dependence on
SimParameters, but this has not been done yet and is not necessary
for any obvious performance reason.

Previous tests showed a 30-40% slowdown for constant pH and alchemy,
regardless of settings. The new code shows only 2-5% slowdown when
omitting alchemical switches, and ~25-30% for 50/50 usage.

Change-Id: I6711f452aa80ae2375ef03595f6e0ed620146674