Expanded alchemical functionality: Drude and CMAP
The following are now compatible with alchemical calculations:
  o Drude Thole damping
  o Crossterms (i.e. CMAP)
  o Drude anisotropic terms

Only the first of these was actually explicitly disabled in prior
releases. Alchemical scaling of crossterms is arguably a very rare
edge case - I can only envision it being used if an entire peptide
or lipid is being annihilated, which is a very ambitious
transformation. However, that in fact was the easiest test case (a
single energy calculation), so it has been implemented.

Thole damping contributes to the electrostatic TI derivative and
FEP energy while the other two go into the bonded term.

Change-Id: I680a176337ed2d920324ec300ca27dec1bc5fddf