Fix CUDA with multiple processes per host
Every process on a host was binding all of the devices on that host,
which is inefficient due to context switching and caused errors when
devices were in compute mode EXCLUSIVE_PROCESS.  Modified DeviceCUDA
to only report devices that are assigned to pes of the calling process.
Also fixed a couple of related bugs.

The ideal configuration for multi-node runs is one process per device.
There is inefficiency and potential strangeness when the number of
devices per node is not a multiple of the number of processes per node,
but it does seem to run.

Change-Id: I310164bd0541afbc2eac1c920462a03fd069e0d4