Fix alchemical coupling + LJcorrection
Integrating alchemical coupling with LJcorrection introduced a
very subtle bug when alchVdwLambdaEnd = 0.0. This is because
the LJcorrection uses a linear combination of the endpoint
corrections, neither of which vanish when alchVdwLambdaEnd = 0.0
(no scaling occurs during the entire simulation). I don't know
who, if anyone is using this, and it should only be a serious
problem when
  a) doing staged VdW/elect decoupling of a ligand
  and
  b) using constant pressure (which is abysmal for performance).

The new behavior just lumps all of the atoms into a single group
when alchVdwLambdaEnd = 0.0, which is actually the same as the
old behavior before integration with alchemy.

Change-Id: I253d76478792a245c1bc8529783304efaa46e54e