Comment on the effects of wrapAll and wrapWater in the UG.
Over time, the wording "alters this behavior" may have become confusing to
novice users, and lead them to think that even the physical trajectory of a
standard MD simulation is changed.

The new explanation is more verbose, but mentions explicitly the use cases
where wrapAll or wrapWater are not recommended.

Change-Id: I68e1ac0734ff0480fb88c0a7f601d7c5140b19eb