Fix bug when aliasing titratable residue names with constant pH
During residue aliasing only the resname was aliased and the
segresidname was not. This broke analysis tools and would have
caused very confusing behavior with residue selection. This is
now fixed here.

If anyone has been using this bug (for all of the last 3 days?),
it would only affect HSD, HSE, and HSP residues which are
otherwise aliased to HIS. A simple sed replacement in stdout,
cphlog, and cphrst files will completely fix the problem going
forward.

Change-Id: Ia2518f625e2110709fd7a8b19190dd7fb660c709