User guide updates for constant-pH MD
Modify documentation to reflect new selection syntax. This just
changes <segid>:<resid> to <segid>:<resid>:<resname> for all
pertinent keywords.

Since the long-hand syntax is a bit verbose, I removed the explicit
repeat to show how optional arguments work. The optional argument
is always strict repetition of the pattern, so this should be self
explanatory. The alternative was an awkward wrapping of the
"Acceptible Values" field that seemed confusing to me.

Change-Id: I086f3e291148ad0dba59dd3a339d62fdafb0de58