Cleanup JSON IO for constant-pH MD
The writing of JSON files (constant pH restarts) is now much
cleaner. There are a few quirks of format that might preclude
using a more general writer, but this was easy enough of a change
without the extra testing that that would require.

Other than a lurking bug in constant-pH specific error reporting,
this outwardly changes nothing that a user will see. It will,
however, make mucking about in the restart format easier as time
goes on.

Change-Id: Ie62dd7cfa42f778078984a73a3e878ef222784a5