Add support for Charm++'s change to use 64 bit ID for load balancing
Based on the value of the CMK_LBID_64BIT compile-time variable,
select either the new 64 bit ID or old 4 int LDObjid to use as
the ID for load balancing, and provide accessor to abstract
access to both types.

This is necessary to support the new 64 bit LB ID interface in Charm++
6.9.0. Once NAMD no longer cares about supporting versions of Charm++
older than 6.9.0, this can be removed and the new interface can be used
unconditionally.

Change-Id: Ice18a9537b694e8e6174cd8118cbc60d59800fe9