Fix HMR bug in constant-pH MD
This patch makes sure that constant-pH MD does not undo mass
repartitioning within titratable residues.

Also, a warning is now issued whenever a proton is automagically
inserted during setup. This is because it is impossible to
on-the-fly repartition mass on those protons and thus the insertions
will not match any pre-existing HMR.

Unfortunately, this means that HMR constant-pH MD will be an
advanced feature for now, since one cannot rely on the auto-setup
within constant pH.

Change-Id: I0243f2e9a781b923a9b5bd83b68b867a0a6fe701