Fix TIP4P water models when using AMBER inputs
The effective deprecation of the "lonepairs" keyword caused some
unexpected behavior when using a parm7 in place of a psf. The main
difference is that lonepairs are now _assumed_ until they are not
found. Since parm7s do not flag lonepairs, the flag is now always
set to false.

Note that TIP4P is a special case of lonepairs, so this is not a
problem. However, the fix does seem a bit counterintuitive, so later
a more involved fix is in order.

The TIP4P lonepairs are treated as a special case.  The rigid bond
constraints routine HomePatch::rattle1old() calls tip4_omrepos()
routine to reposition the lonepair site after the atom positions
are updated on each water molecule.

Change-Id: I91f430f79c810ad0ff13999de4cdcd453f6171ff