PME CUDA recalculate self-energy term
Modify PME CUDA implementation to always recalculate the self-energy term.
This solves the incompatibility with REST2, in which PME CUDA was getting
the energy wrong after a charge rescaling.

Also greatly simplifies code by submitting self energy to reduction system
where it is calculated rather than separately reducing it to a pencil.

Change-Id: I05d8d959004b5e7ed8a274f41936a9400a4a7e1d