Fix hypothetical PME race condition
In PmeZPencil::send_all_ungrid using CkLoop if the pencil's
home PE does not wait for the helper PEs to finish sending
it could in theory receive incoming data for the next step
and overwrite the charge/potential grid before it was sent.
This would be very unlikely, but it was triggering a recently
added bug check because hasData was being reset to 0 by the
host PE before being checked by the helper PEs.

Change-Id: I4ec7b22ae4f3035c1328498a2cc8f5c4665171c5