Simplify microscopic inherent pKa calculations
This revision doesn't change the function of the code, but puts the
implementation more inline with the equations in the published paper.

Briefly, the multiplicity of states must be taken into account when
computing transition probabilities based on the inherent pKa. This
is most clearly written as a correction to the macroscopic pKa input
in the form of a log-ratio of multiplicity factors for the current
and trial states. This was previously split into two corrections,
which was technically correct, but obscured the ratio component.
This also reduces the number of logarithms computed.

Additional change:

The input pattern for a two-site degenerate diprotic acid has been
implemented. This was just a convenient feature for testing and does
not match any published force field parameters.

Change-Id: If3ddb84fc6271b2f261055c82ee98f055ed49310