Warn that extraBonds angles are cosine-based and allow harmonic
The "normal" flag for extraBonds angles was left uninitialized
when the cosAngles feature was implemented, which tends to
result in extraBonds angles being treated as cosine-based.

Since workflows may unknowingly depend on the current behavior
warn the user and add extraBondsCosAngles option to change it.
Also fix bad comment that reversed meaning of normal flag.

Change-Id: I1a300d6b12654c676c91f15236b135ef3e5fc4e3