Add alchIDWSFreq fallback to outputEnergies
If alchOutFreq = 0, fallback to alchIDWSFreq = outputEnergies. This
enables IDWS output to stdout without sending any output to
alchOutFile.

Also, because the value of alchLambda2 is step-dependent (not event
driven), the use of two IDWS output streams may be inherently
incompatible if the output frequencies are different (outputEnergies
must be an odd multiple of alchOutFreq). Users can still use whatever
settings they would like, but there is a warning issued when
alchOutFreq > 0 and alchOutFreq != outputEnergies.

Minor change:

alchIDWSfreq --> alchIDWSFreq, to match camel casing convention
throughout the rest of the code

Change-Id: I5d6611b6eb9374cd1cc97e1e87169e4beba83322