Adding error message for large patches on nbond kernels
We had an 'invalid argument' error when large patches (>11k atoms) were
accounted for on CudaTileList::buildTileLists. This happened because
the amount of shMem allocated inside the tilelist kernels
is dependant on the number of atoms in the largest patch.

I've added a query to get the maximum shared memory available per block.
We can compare that against the necessary shared memory to launch the
tilelist kernels, and if those launches are about to go wrong, NAMD prints
a nice message to warn the user about the patches sizes.

Change-Id: Ie35b40d83f87b06ec33cd60b0a1936cec18b6114