Update running CUDA and performance tuning
Moved performance tuning chapter to later part of user guide.
Updated to discuss NAMD performance tuning concepts.

Updated section on running CUDA for modern GPUs, added explanation
of what is and is not accelerated, what is and is not disabled by
CUDA-accelerated builds, and added some keywords.

Updated copyright year on user guide.

Change-Id: I21cda6ef7bd25fde53274d0733637b645645c2f8