Improve energy conservation of vdwForceSwitching/LJcorrection
The existing implementation for vdwForceSwitching + LJcorrection
introduces a discontinuity into the potential (but not the force)
in order to make the long-range correction more well-defined.
While this discontinuity is very small for normal cutoffs, it
can cause detectable issues at moderate lengths (~8 A).

This revision reverts to the old vdwForceSwitching behavior and
modifies the long-range correction to include a "core" correction
term based on the same approximation -- this is theoretically
dubious, but works quite well in practice.

On a practical note, this change greatly simplifies the code
because vdwForceSwitching behavior is now uniform. Supporting
multiple branches in the alchemical code will now be substantially
easier.

Change-Id: Ia53bc80f79e1976f2f0ca62f2f0d6988a2fe960a