Update behavior for alchemical bonded scaling
Molecule::get_fep_bonded_type now explicitly catches bonded terms
involving alchemical atoms that should NOT be scaled and assigns
them to group 0 (as opposed to 1 or 2). This may have been a quiet
bug for the case alchBondLambdaEnd > 0.0 and alchBondDecouple off,
neither of which are default and are not often (or never?) used.

Change-Id: I404703eadd30951270ff2a453e8cdff31c835bdc