Use NAMD_die when exiting due to errors
Fixes problem with job chains going on after crashes, because the scheduler is
unaware of the error condition.

This simple version should work with any version of Charm++. I am not sure what
the current logic tries to achieve, as BackEnd::exit() is called if a single partition
is present.

A more sophisticated version might take advantage of this recent change:
https://charm.cs.illinois.edu/gerrit/c/charm/+/4219
which would allow BackEnd::exit() to accept an exit code and pass it along
to charm++.

Change-Id: I75c6361947ee383b250d6d6e3faf3bd5769c5096