Experimental code for serial expanded ensemble in pH space
pH replica exchange is a well-established technique in the literature, but
is difficult to implement efficiencly in the inherently asynchronous paradigm
of neMD/MC constant-pH. This code implements some well-know serial variants of
pH replica exchange and creates a simple interface for trying other versions.
They have been tested and are theoretically correct, but require considerable
calibration and expert knowledge to produce meaningful results. Thus they
ought to be considered experimental.

The code changes do not actually plug in to the main cphRun function and thus
do not actually provide new features to namdcph. They are being added so that
newer developers can try new ideas in the future.

Change-Id: Ie5fff39caa401563c5ad7392a1d88fbd6ef0159e