Merge profiling from devel branch
Includes NVIDIA NVTX, Charm++ Projections tracing user events, and
detailed in-house timer collections with optional histogramming.
The profiling here is toward understanding the CPU performance of the
numerical integration parts of the code in Sequencer and HomePatch.

To build with NVIDIA NVTX:
- define NAMD_NVTX_ENABLED (compiler flag -DNAMD_NVTX_ENABLED)
- link to libnvToolsExt (Make.config use "EXTRALINKLIBS = -lnvToolsExt")

The Charm++ tracing user events requires verion 6.9.0 of Charm++.
To build with Charm++ Projections tracing user events:
- build Charm++ with "--with-production --enable-tracing"
- define CMK_TRACE_ENABLED (compiler flag -DCMK_TRACE_ENABLED)
- define NAMD_CMK_TRACE_ENALBED (compiler flag -DNAMD_CMK_TRACE_ENABLED)
- build NAMD using "projections" target, e.g. "make projections"

To build with timer collections:
- define TIMER_COLLECTION (compiler flag -DTIMER_COLLECTION)
- optionally define TIMER_HISTOGRAM (compiler flag -DTIMER_HISTOGRAM)

Change-Id: Ia77ba9d4a6ca4ae81deb65103a8ee67ba74d485a