Initial support for Python3
This ports the Python module "tcl" to work with Python3 interpreters.
It might need some refinement, eg. in the handling of Unicode strings.
Right now it assumes only ASCII strings are passed.

Processing exceptions from Python does not seem to work with Python3 as
is. Added a null pointer test to avoid a segfault.

Change-Id: I8bdaddd5d02970ceffb4fbacae6369915b8a8612