Fix signed integer overflow in RNG
Change random number generator to use unsigned integer to avoid undefined
overflow behavior of signed integer, which was causing a hang for g++
builds.  The g++ compiler sets a signed integer overflow to its maximum
value, but this was causing an infinite loop in Random.h lines 124-128 due
to the inability to advance rand48_seed in Random::skip().

Change-Id: I8ab0445277cfdac79a99e3217aeec4d650204eb4