Split off dependencies for individual libraries into separate files
Make.depends can easily get out of date due to concurrent changes in multiple
parts of the codebase.  This is further complicated for source files that are
developed in separate repositories, like Colvars or Lepton.

This change isolates the dependencies of files that are tracked by separate
repositories, by updating the re-generation Makefile recipe and including an up
to date copy of Make.depends and its subsidiaries.

Change-Id: Ib45c1e3e27da6dcbc77a12b22fd8e88186958883