Clarify error for CUDA with MPI/non-SMP Charm++
The error message printed by the config script when attempting
to use CUDA with an inappropriate MPI-based or non-threaded
Charm++ arch was not clear about the exact reason for the error.
This change prints the reason as well as the offending architecture.

Also updates suggestion of old ibverbs-smp to new ucx-smp.

Change-Id: I0c2952c6fd7200462ccc170c6495728868ff5303