Let ScriptTcl call externally-defined Tcl wrapper for Colvars
This eliminates the need for Colvars to delete ScriptTcl's own "cv" Tcl
command in order to replace it.

It also fixes an error when the "cv" command needs the new implementation but
this hasn't been created yet (i.e. when the NAMD Tcl commands are suspended
until startup).

A future update for Colvars will keep using this updated function rather than
replace it, so that the semantics of ScriptTcl execution are overall
respected.

Change-Id: I7e0a5249bc83ed20db49c5207fde9b6563fec82c