Allow using GridForce maps in GlobalMaster (backend for map-based colvars)
This commit introduces the use of GridForce maps in GlobalMaster, which is
leveraged by the Colvars module to implement new collective variable types.
It is submitted to master because this code has been developed and tested
there, but may be easily rebased onto devel if needed.

When mGridForceScale is a triplet of zeros, the corresponding map does not
produce any gridForces potential and its computation should be skipped to save
time (see condition added in ComputeGridForce.C).

Additionally, the same condition is now also detected inside GlobalMaster and
ComputeGlobal, to allow using the corresponding map for a global computation
(see following commits).  A list of numeric indices of GridForce maps
requested for global computation is collected by GlobalMasterServer from
individual GlobalMaster objects.  The list is then shared with ComputeGlobal
instances on the nodes, so that the corresponding maps are computed on each
patch and communicated back to GlobalMasterServer.

To properly collect all contributions, GlobalMasterServer must not only count
atoms and groups if any are requested (as it did before), but also count
patches when one or more GridForce maps are requested.  This adds the
synchronization barrier necessary for global forces processing.  Thanks to Jim
Phillips for pointing out this requirement.

A corresponding change in Colvars has been introduced to allow users access to
this new feature, and will be submitted separately.

Things that may be improved:

- Instead of mGridForceScale == 0.0, a SimParameter option may be added
  instead.  I went with the former only to minimize changes to existing code.

- The synchronization barrier associated to counting patches instead of atoms
  may become unnecessarily long for large systems, especially if most of the
  patches do not contribute to the maps (e.g. studying the hydration of a
  small cavity in a very large proten complex).  Further optimization may be
  achieved by collecting all patches and counting those that have a
  significant contribution every stepsPerCycle steps, and collecting just
  those patches at the other steps.

Change-Id: If0d0456abbb13d76d474ae2fc7d50948b919e6b2