Fix Psfgen documentation
Fixed a formatting error in the Psfgen documentation, in which an
unterminated \tt used to offset a hyperlink accidentally changed the
fonts for the rest of the document. Changed the hyperlinks in the
Psfgen section to use the hyperref package \url{ } so that they
correctly link and match the others.

Also fixed the original Psfgen documentation in the VMD plugins tree,
so that we don't undo these changes with a future import of Psfgen.

Change-Id: Ida89c0be62daa63e2e8d8bf7a133f96c1589f22a